requirements.txt
Django==4.2.21
djangorestframework==3.16.0
django-filter==25.1
dj-database-url==2.3.0
psycopg2-binary~=2.9.3
